Open GuiRecipe: normalizeItemStack
